requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
python-dotenv==1.1.1
pytz==2025.2
pydantic==2.10.5